// src/app/models/schedule.ts
export class Schedule {
    scheduleId?: number;
    airlineId?: number;
    flightId?: number;
    arrivalTime?: Date;
    departureTime?: Date;
    startFrom?: string;
    destination?: string;
    scheduledDays?: string;
    ticketCostForBusiness?: number;
    ticketCostForNonBusiness?: number;
    mealPreferences?: string;
    createdDate?: Date;
    createdBy?: number;
    modifiedDate?: Date;
    modifiedBy?: number;
    Isroundtrip?:string;

}